Adds new screen about joining a project. Adds a new screen for starting a project. Navigates to those screens. Make svgs color agnostic.
